Moved Es6Rewriter into its own file `src/es6-rewriter.ts`
 - PR feedback https://github.com/Polymer/polymer-bundler/pull/637#discussion_r172717310
